requirements.txt
crewai>=0.28.0
crewai-tools>=0.8.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
pyyaml>=6.0